feat: adds user id validation in NotificationService

+++ Service/src/main/java/ru/isaev/Service/NotificationService/NotificationService.java
@@ -5,10 +5,13 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import ru.isaev.Domain.Notifications.Notification;
 import ru.isaev.Domain.Products.Product;
+import ru.isaev.Domain.Users.Roles;
 import ru.isaev.Domain.Users.User;
 import ru.isaev.Repo.NotificationRepo;
 import ru.isaev.Repo.UserRepo;
 import ru.isaev.Service.Security.MyUserDetails;
+import ru.isaev.Service.Utilities.Exceptions.NotYourNotificationException;
+import ru.isaev.Service.Utilities.Exceptions.NotYourProductException;
 import ru.isaev.Service.Utilities.Exceptions.UserNotFoundException;
 import ru.isaev.Service.Utilities.NotificationsTimestampComparator;
 
@@ -20,10 +23,15 @@ public class NotificationService implements INotificationService {
 
     private final NotificationRepo notificationRepo;
 
+    private final User currentUser;
+
     @Autowired
     public NotificationService(UserRepo userRepo, NotificationRepo notificationRepo) {
         this.userRepo = userRepo;
         this.notificationRepo = notificationRepo;
+
+        MyUserDetails currentPrincipal = (MyUserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        currentUser = currentPrincipal.getUser();
     }
 
     @Override
@@ -32,14 +40,17 @@ public class NotificationService implements INotificationService {
                 () -> new UserNotFoundException("Not found notification with id= " + id)
         );
 
+        if (!currentUser.getId().equals(notification.getUserId()) && !currentUser.getRole().equals(Roles.ROLE_ADMIN))
+            throw new NotYourNotificationException("You do not have access to notification with id= " + notification.getId());
+
         return notification;
     }
     @Override
-    public List<Notification> getAllNotificationsOfCurrentUser() {
-        MyUserDetails currentPrincipal = (MyUserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        User currentUser = currentPrincipal.getUser();
+    public List<Notification> getAllNotificationsOfUserById(Long userId) {
+        if (!currentUser.getId().equals(userId) && !currentUser.getRole().equals(Roles.ROLE_ADMIN))
+            throw new NotYourNotificationException("You do not have access to notifications of user with id = " + userId);
 
-        List<Notification> notificationsList = notificationRepo.findByUserId(currentUser.getId());
+        List<Notification> notificationsList = notificationRepo.findAllByUserId(currentUser.getId());
         notificationsList.sort(new NotificationsTimestampComparator());
         return notificationsList;
     }
@@ -62,6 +73,13 @@ public class NotificationService implements INotificationService {
 
     @Override
     public void deleteNotificationById(Long notificationId) {
+        Notification notification = notificationRepo.findById(notificationId).orElseThrow(
+                () -> new UserNotFoundException("Not found notification with id= " + notificationId)
+        );
+
+        if (!currentUser.getId().equals(notification.getUserId()) && !currentUser.getRole().equals(Roles.ROLE_ADMIN))
+            throw new NotYourNotificationException("You do not have access to notification with id= " + notification.getId());
+
         notificationRepo.deleteById(notificationId);
     }
 }
